Simplified code, created CopyAdapter and PdoAdapter
