+++ composer.json
@@ -25,7 +25,8 @@
     "require-dev": {
         "keboola/coding-standard": "^8.0",
         "phpstan/phpstan-shim": "^0.11.15",
-        "phpunit/phpunit": "^8.3"
+        "phpunit/phpunit": "^8.3",
+        "ext-json": "*"
     },
     "autoload": {
         "psr-4": {
