+++ src/Exception/CopyAdapterException.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Keboola\DbExtractor\Exception;
+
+class CopyAdapterException extends ApplicationException
+{
+
+}
