+++ tests/phpunit/IncrementalFetchingTest.php
@@ -493,7 +493,9 @@ class IncrementalFetchingTest extends BaseTest
         $newResult = ($this->createApplication($config, $result['state'], $logger))->run();
 
         $this->assertTrue($handler->hasWarningThatContains('Forcing extractor to use PDO fallback fetching'));
-        $this->assertTrue($handler->hasInfoThatContains('Executing query via PDO'));
+        $this->assertTrue($handler->hasInfoThatContains(
+            "Executing query 'in.c-main.auto-increment-timestamp' via PDO ..."
+        ));
         //check that output state contains expected information
         $this->assertArrayHasKey('state', $newResult);
         $this->assertArrayHasKey('lastFetchedRow', $newResult['state']);
